Add files via upload
Here You can find the code of the correct K-Means and in line with the paper. Unfortunately, we realized only after submitting the code that when we created the final file, we had taken the wrong code from the K-Means (in fact, a completely different result is written in the paper than the result in the code). We apologize for the error. 

(Obviously, the rest of the code is the same as the original)